requirements.txt
mysql-connector-python==8.0.13
mysql-connector-python-dd==2.0.2
pandas==0.23.4
requests==2.9.1
scipy==0.17.0
urllib3==1.24.1

